requirements.txt
numpy
scipy
soundfile
sounddevice
PyYAML
onnxruntime
optimum[onnxruntime]
transformers
streamlit
